+++ traffic_portal/test/end_to_end/login/login-spec.js
@@ -20,7 +20,8 @@
 var cfunc = require('../common/commonFunctions.js');
 
 describe('Traffic Portal Login Test Suite', function() {
-	var commonFunctions = new cfunc();
+	const commonFunctions = new cfunc();
+	// browser.get(browser.baseUrl);
 
 	beforeEach(function() {
 		browser.get(browser.baseUrl + '/#!/cdns');